requirements.txt
asn1crypto==1.1.0
cffi==1.13.0
cryptography==2.7
Django==2.2.5
django-filters==0.2.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
